requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1
anthropic==0.30.1
tenacity==8.5.0